Implemented Fallback for Current Location
Added fallback for current location. It follows the same methodology as other options (allowing for per-computer settings).

This also supports the ‘home’ and ‘work’ modifiers. Setting to ‘here’ as a fallback is bad and has been accounted for.

Also took the opportunity to do some code cleanup. A lot of functions were moved into the functions file to eliminate duplicate code as part of implementing this feature.

Signed-off-by: Stuart Ryan <stuart@ryan.email>